+++ src/js/api/order.js
@@ -644,6 +644,47 @@ ripe.Ripe.prototype.cancelOrderP = function(number, options) {
 };
 
 /**
+ * Changes the priority of an order.
+ *
+ * @param {Number} number The number of the order to change the priority.
+ * @param {Number} priority The new priority.
+ * @param {Object} options An object of options to configure the request.
+ * @param {Function} callback Function with the result of the request.
+ * @returns {XMLHttpRequest} The XMLHttpRequest instance of the API request.
+ */
+ripe.Ripe.prototype.setPriority = function(number, priority, options, callback) {
+    callback = typeof options === "function" ? options : callback;
+    options = typeof options === "function" || options === undefined ? {} : options;
+    const url = `${this.url}orders/${number}/priority`;
+    options = Object.assign(options, {
+        url: url,
+        method: "PUT",
+        auth: true,
+        params: {
+            priority: priority
+        }
+    });
+    options = this._build(options);
+    return this._cacheURL(options.url, options, callback);
+};
+
+/**
+ * Changes the priority of an order.
+ *
+ * @param {Number} number The number of the order to change the priority.
+ * @param {Number} priority The new priority.
+ * @param {Object} options An object of options to configure the request.
+ * @returns {Promise} The result of the order priority change.
+ */
+ripe.Ripe.prototype.setPriorityP = function(number, priority, options) {
+    return new Promise((resolve, reject) => {
+        this.setPriority(number, priority, options, (result, isValid, request) => {
+            isValid ? resolve(result) : reject(new ripe.RemoteError(request, null, result));
+        });
+    });
+};
+
+/**
  * Imports a production order to RIPE Core.
  *
  * @param {Number} ffOrderId The e-commerce order identifier.