+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `getDefaultsP` - the promise based version of `getDefaults`
+* `changePriority` Order API method
 * `tenancyAccountMe` endpoint
 
 ### Changed